Convert some uses of printf to cout to avoid format warnings
The printf format specifier for uint64_t may be %ul or %ull depending on
the platform. There is a portable macro (PRIu64) that one can use
instead, e.g.,

  printf("%" PRIu64 " the rest", val);

This approach looks gruesome to me.

Alternatively, we can use cout style printing instead. It is slightly
slower than printf but if we avoid endl, which flushes, the performance
is pretty similar [1].

In our codebase, cout is slightly preferred anyway:

  $ find . -name '*.cpp' -o -name '*.h' | xargs grep printf | wc -l
    2280
  $ find . -name '*.cpp' -o -name '*.h' | xargs grep cout | wc -l
    4611

One day, C++20 will become widely available and we can use std::format,
which is the best of both worlds.

[1] https://stackoverflow.com/questions/2872543