Fix build on Arch Linux
The newer boost libraries 1.72 on Arch Linux only provide the
multithreaded variants. This causes a CMake configuration failure
because we explicitly set(BOOST_USE_MULTITHREADED OFF).

The most forward looking fix would be to remove the multithreaded
constraint but I'm not sure of the performance implications thereof. For
the time being to unblock the CI, just pin the boost library to a
previous version that does provide non-multithreaded variants.