[scripts] check cpp files as well
I was running this script on my code and noticed that I appeared to have
written perfect *.cpp files. That seemed very unlikely; find wasn't
printing out cpp file names. Not sure if my version of find is different
than the mac version or if this is a general bug.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>