Fix wakeup overhead test on two-threaded machines
The following assertion in ThreadPool::runInternal fails on machines
with just two threads:

  assert(!masterFastmode || masterFastmode == num);

This is due to the fact that runInternal expects num/masterFastmode to
be clamped to the number of usable threads (i.e., active threads less
reserved threads), but ThreadPool::burnPower sets masterFastmode
without any validation.

This assertion does not fail with machines with more than two threads
because in the wakeup test activeThreads is always set to 2; so
getMaxUsableThreads will always be greater than activeThreads, and the
edge case in burnPower is never exercised.