ci: Upload debugging data on CI failure
It takes about 5-10 minutes for the uploaded dump data to appear as a
Github artifact.

The build directory is about 2GB. I spent some time figuring out if we
can reduce the size but core dumps are about 2GB each as well. And going
down the path of filtering files may create situations where we drop too
much and we defeat the point of post-mortem debugging.

Test and build output is already streamed to the terminal, so the dumps
should only be used in rarer cases.