ci: Avoid platform-specific optimizations
The GitHub build farm includes a mixture of processor architectures. In
particular, some machines support AVX-512 and some don't. This interacts
badly with ccache, which caches object files based on compiler
invocation flags and does not include any information about the host
platform: "-march=native" on one machine means something else on
another.

In particular, this resulted in a test failure in floatingPointErrors:

  SIGILL

  => std::mersenne_twister_engine<...>&)+109>:    vcvtusi2sd %rax,%xmm0,%xmm0