Test and build with clang
The original impetus was to get coverage for an issue using
setjmp/longjmp which apparently is triggered with -march=native and
clang for some definitions of native.  In my brief testing with the
Github CI, I couldn't get failures to happen with or without
march=native; though I always observed failures locally in Debug mode.

So we can't get the coverage we want, but I think the larger question of
some cross-compiler verification is still relevant. I chose to test
clang and Release in particular because that configuration is the most
likely to be "different" (so more like a canary than anything else).