ci: Enable testing
Two non-obvious things about CI testing in Travis and CircleCI (and
relatedly Docker testing infrastructure):

MPI doesn't like being run as root. An example of the heated debate why:
[1].

CircleCI has some issue reading from the home(?) directory (which it
calls `working_directory`) when you are not root. The symptom was an
inability to read the /root/project directory as a non-root user. In my
local testing in Docker, this was never an issue. I worked around this
by moving the build directory to /tmp/.

[1] https://github.com/open-mpi/ompi/issues/4451