Remove gtuple
icc fails to compile gtuple due to errors like the following:

  [...]gtuple.h(77): error: pack expansion does not make use of any argument packs
      using append = integer_seq<T, Is..., I>;
                                      ^
            detected during instantiation of class "galois::integer_seq<T, Is...> [with T=int, Is=<>]" at line 104

Similar errors are reported by icc when it compiles <tuple> from the GCC
9 STL library. Currently, icc 12 only supports GCC 8 as a backend.

So, it appears that there are some C++17 features that still trip up
icc. As a workaround, reformulate gtuple in terms of standard C++17 STL
features. Since, gtuple itself was designed to provide transitional
support for integer sequences and related functionality in mixed C++11
codebases, this gives us an opportunity to remove custom code.