Fix race when PerThreadStorage reaches limit
The bump allocation does not validate that the reserved offset is
actually within the allocated space. This will cause undefined behavior
if multiple allocations are made as the allocator reaches capacity.

Along the way, switch over to <atomic> from __sync intrinsics. I tried
to preserve the same memory consistency as before on the bump allocation
path. The memory consistency is stronger than before along the other
paths (including deallocation) because the default behavior of <atomic>
is sequential consistency, but I felt that simplicity is warranted here.