Standardize uses of GALOIS_DIE
Error messages should be user actionable if possible. This means that
they should expressed in terms of user-facing abstractions (e.g., files,
command line arguments) and phrased in such a way as to indicate what
the remedy should be.

Along the way, give errors with the same causes the same message, adopt
an initial lower-case for messages, and remove excess developer-facing
context. The line number printed by GALOIS_DIE should be sufficient to
identify the particular GALOIS_DIE statement.