Add RPC services for basic manager-daemon logic.
The manager must be started at some node first. Then when a spawn daemon is
started, it sends a `RegisterDaemon` RPC to the manager to register itself.
The manager sends back a `SpawnWorker` RPC to ask the daemon to
initialize the API server pool. As the acknowledgment, the daemon
replies with a vector of the spawned API servers' information.

Currently the only daemon and API server information is their ports.