MAP_POPULATE local files at least
There is a question as to whether MAP_POPULATE is wise for all file
descriptors but for descriptors backed by the local file system it seems
reasonable pre-fault the file in and avoid contending on the process
address space lock during threaded execution.

This also matches the behavior of FileGraph.cpp and this is a precursor
to moving more file operations out of libgalois and into libtsuba.