Python interface for Galois data structures and Jupyter/Cython Galois usage (#87)
Add support for Python level access to Galois data structures (using a Jinja2 generated Cython file) and updates the python examples to be usage from Python directly. This enables the use of Jupyter notebooks to use Galois via the Cython support in Jupyter. There are still some issues with the installation and launch process, so here is some documentation.

To setup Jupyter to work with Galois:
* Install Jupyter
* Build/Install Galois Python (either with `setup.py install` or a wheel or just `setup.py build` and then setting environment variables correctly)
* Launch Jupyter using: `python3 -m galois.launcher jupyter notebook`. The module `galois.launcher` will setup the environment variables as needed by Galois and Cython.
* Put `%load_ext cython` in your notebook.
* Add `# distutils: extra_link_args=["-lgalois_shmem"]` to each `%%cython` block in your notebook that uses Galois.

Look in `python/examples/jupyter` for examples of notebooks which work in this environment. The files to load will need to be changed to make them work. (An autodownloader would be a good idea to enable better examples in the future.)