libsupport: Add header for platform specific information
I guess MAP_POPULATE is not POSIX; it breaks the build on Mac OS. There
is a whole bunch of ifdefs scattered in places like
`libgalois/src/PageAlloc.cpp` and `libgalois/src/FileGrpah.cpp` to
protect things like this.

I think having a place like Platform.h were we can keep track of these
things in a central location is a good idea. So far I've done the bare
minimum so that this will compile on Mac.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>