tsuba: subsume parquet parsing code
Split PropertyFileGraph into a graph engine part (called
PropertyFileGraph) and a storage engine part (called RDG). RDG's are
simple structs embedded in PropertyFileGraph that are populated by
tsuba, conversely the graph engine can fill an RDG struct and pass it
back to the storage engine to be written out.

Graph names are the same as the name for the meta-data file right now.

The metadata struct is now opaque to the graph engine and controlled by
tsuba so that we can do things like enforce capabilities down the line.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>