QueryTest app uses PFGs; removed DBGraph
DBGraph is now an unnecessary wrapper around AttributedGraph as we are
not loading plain unattributed graphs for this anymore. querytest now
only supports PFGs and serialized boost graphs (backward compatibility).
Commit also includes a spacing change in AttributedGraph. querytest
tests are disabled for the time being as they need to be updated to
reflect the new PFG usage.