Return value parsing from Cypher query
Compiler was not actually parsing values of return query before. Support
now added for it to parse those queries and construct in-memory
representations of each return predicate for use by the engine later in
returning what needs to be returned after a query is run. Currently does
not support arbitrary property nesting nor function calls other than the
count function.

[REWRITTEN COMMIT: incomplete subtree]