Avoid separable compilation by default
When separate compilation is on, cmake will invoke nvcc with the
-dc/--device-c option. This will generate relocatable device code for
each translation unit, which will be resolved when linking an
executable. This matches how CPU code is normally compiled and linked,
but it adds a small performance penalty because functions need to be
relocatable (e.g., standard caller/callee saved registers).
Traditionally, CUDA code is compiled and linked in one step, which
simplifies optimizations through function boundaries.

The CMake configuration before this change, for instance, used separable
compilation for the barneshut and delaunyrefinement applications and it
was commonly on for other GPU applications [1].

It turns out that no current applications require separate compilation
to build, and one application, matrixcompletion, which previously was
built without separable compilation actually requires it; otherwise,
nvcc generates incorrect code. Digging further is a job for another day.

AvA currently requires non-separable builds.

Given all the above, it seems reasonable to bias towards non-separable
builds unless there is a strong reason otherwise.

[1] https://github.com/IntelligentSoftwareSystems/Galois/pull/283

[REWRITTEN COMMIT: incomplete subtree]