build: Honor BUILD_TESTING
BUILD_TESTING is a standard CMake configuration option that is used by
callers to indicate whether a CMake project should build tests.

The standard idiom is if BUILD_TESTING=ON and we are the only project,
build tests. If we are included in another project, never build tests.
BUILD_TESTING is ON by default.

The main wrinkle is how to treat lonestar applications. When
BUILD_TESTING=OFF, we only build libraries, which is sensible behavior
for someone interested in packaging this repo as a library. When
BUILD_TESTING=ON, we build the applications and tests. There is no
option to build one but not the other.

This seems permissible for now.

Building tests without applications doesn't make much sense because the
majority of tests simply run applications in various ways.

Conversely, the majority of the build time is building applications.
Building tests themselves takes a fraction of the total time.

[REWRITTEN COMMIT: incomplete subtree]