Return edge labels/attributes via Cypher
Users can now write return <x> where x is an edge variable, and the
engine will print the data as requested. This PR also contains
miscellaneous additions to support this feature, including a function
to get a string representing the labels of an edge, changes to the
LabeledGraph's binary search function to return the found edge iterator,
and a findEdge function in LabeledGraph that, unlike the existing
isConnected function, will return the found edge iterator.

Note that return attributes printed are currently incorrect due to an
issue found while adding this feature in; see issue #230 for context. A
PR to fix this is coming soon.

[REWRITTEN COMMIT: incomplete subtree]