Add Katana to Katana transformers
The Parquet representation of a boolean array is strictly a bitmap. This
prohibits useful storage optimizations like run-length encoding and
sparse row groups. Also the libarrow implementation of parsing Parquet
bitmaps is time-consuming as well.

Introduce a sparsification transform that converts a boolean array to a
byte array. This reduces storage significantly for properties with many
zero values. It also improves load time.

Additionally, add a timestamp parser.

[REWRITTEN COMMIT: incomplete subtree]