Return to large row group chunks
The chunk_size parameter to WriteTable actually controls the size of
pages within a row group.  In testing, reverting to the original chunk
size was about twice as fast for loading than using the small chunk
size. Reducing the chunk size was introduced to improve the memory
behavior and performance while converting large files, but adjusting it
monolithically seems inappropriate.

[REWRITTEN COMMIT: incomplete subtree]