Cypher: array access account for sorted edges
Before this commit, arrow array accesses for edges were incorrect as the
underlying CSR graph is sorted for efficiency while the arrow arrays are
not sorted along with them. This causes references to a particular edge
index to be incorrect. This commit adds a mapping from original edge
index to new edge index that's used when accessing the arrow arrays; it
is constructed by sorting a 1-to-1 mapping in the same manner as the CSR
is sorted.

The original plan was to sort all of the arrow arrays to avoid a layer
of indirection for each edge property access, but complications there
included creating a new arrow chunked array for the sorted array which
included having to deal with null elements during the sort. This is
not to say this isn't doable; in fact, the original to new edge mapping
created in this commit can be used to facilitate the creation of new
arrow arrays for this purpose. That can be done in another PR.

Resolves #230

[REWRITTEN COMMIT: incomplete subtree]