Decouple threads from partitions in dist tests
Distributed tests will now always use up to 4 processes and a maximum
of 8 threads per process. This prevents the need for specifying
NUM_TEST_THREADS in order to spawn tests with more than 1 process and
makes it so multi-process tests always occur regardless of machine
configuration. NUM_TEST_THREADS no longer needs to be set in the
cpp.yaml file as a result.

[REWRITTEN COMMIT: incomplete subtree]