Cypher compiler: return DISTINCT, LIMIT, SKIP
The Cypher compiler now recognizes 3 modifiers for the return statement:
distinct, limit, and skip. These are saved to a struct and can be
accessed by the user via a public function to determine the manner in
which the query matches should be returned. The Cypher compiler tool now
prints this info for debugging purposes as well.

The functions/structs added to support this were designed with the WITH
Cypher clause in mind: WITH is very similar in structure to RETURN as it
shares many of the same modifiers (including the 3 added by this
commit).

[REWRITTEN COMMIT: incomplete subtree]