Compiler: add struct to hold common func(var.prop)
<func>varname<.prop> is a very common pattern in many cypher queries.
This commit makes it its own structure as many constructs will make use
of it as new features are added to the compiler.

This also makes the return struct slighly more general (e.g. a boolean
solely for tracking "count" doesn't need to exist anymore as we now
store function names).

Also flipped on printResults for the querytest test to make sure
printResults doesn't cause crashes.

[REWRITTEN COMMIT: incomplete subtree]