Pick one pattern for iterating over per-thread containers
There are two general patterns in the code base for accessing per-thread
containers: TwoLevelIterator.h and TwoLevelIteratorA.h.

TwoLevelIterator.h wraps several standard STL types like std::vector
(via gstl::Vector) into per-thread variants (like PerThreadVector).

TwoLevelIteratorA.h provides a general multi-level iterator that wraps
any container of containers and produces a iteration space of elements
of the inner container.

The intent of TwoLevelIterator.h is to provide a curated interface that
hides the fact that a single logical collection is implemented by N
per-thread collections (and provides sensible defaults for allocators).
In practice, not many applications use this interface and the additional
level of class wrapping is a source of drift between the wrapped class
and the exposed interface.

TwoLevelIteratorA.h adopts a more general strategy. Any collection of
collections can be the source of elements for parallel iteration. The
downside is that adaptation of iterators in general may lose out on
specialized iteration strategies. Also, TwoLevelIteratorA could benefit
from some straightforward performance improvements. But if the main use
of this adaptor is to facilitate copying into worklists, only forward
iteration needs to be fast.

In any event, I think is better to have one way to do things rather than
two.

[REWRITTEN COMMIT: incomplete subtree]