Avoid perfect forwarding when object lifetimes do not match
Forwarding references, which is when an argument is a non-cv qualified
type declared as a template parameter, e.g.,

  template<typename Arg>
  Fn(Arg&& arg)

allow template parameters to match reference types (T&) where normally,
they are decayed to the underlying non-reference type (T).

Preserving references is useful for passing function arguments because
we can preserve the intent of the caller, but in one case in libgalois,
`galois::wl`, forwarding references are used to determine the type of
fresh objects. Since these new objects have a separate lifetime from the
arguments of the function that created them, this creates the
possibility for dangling references, e.g., when the object type is
deduced to be a reference but the referent is stack allocated.

When caller and callee lifetimes do not coincide, we must make a copy.
There are two approaches:

1. Copy and move

  Fn(Arg arg): arg_(std::move(arg)) {}

2. SFINAE undesirable constructors

  template<typename U, std::enable_if_t<std::is_convertable_v<U, T>>* = nullptr>
  Fn(U&& arg): arg_(std::forward<U>(arg)) {}

Copy and move introduces an extra move for copyable types and an extra
copy for types that are copyable but not movable. SFINAE requires
advanced C++ knowledge to understand and, in our particular case of
tuple types, will require some more meta programming to apply to each
tuple element in the argument and in final storage object.

Since galois::wl exists just to temporarily store arguments to an
indexing function, I opted for the simpler and more obviously correct
copy and move strategy.

I checked the rest of libgalois and all other uses of std::forward are
either strictly forwarding or create a temporary that does not live
outside the function call.

[REWRITTEN COMMIT: incomplete subtree]