CompilerThing->QueryProperty, remove QueryResult
This commit renames CompilerThing to something more sensible and
also adds an alias field to it so that CompilerQueryResult can be
removed completely from the system. QueryProperty takes the role of it,
and the type declarations throughout the system are changed accordingly.
The issue with this is that for non-return values you'll have a dead
field (alias) since it only makes sense in RETURN or WITH, but the
tradeoff is one less struct to deal with.

[REWRITTEN COMMIT: incomplete subtree]