tsuba: Make tsuba compatible with localstack
Localstack requires path-style URIs, enable these when an alternate
endpoint is provided. S3 has deprecated path-style URIs so always using
them is not a good choice.

Localstack is tough to integrate with current CI for other reasons
(uses docker with is not supported on MacOS builds). But supporting
path-style URIs is via this strangely named constructor argument is
subtle and I would hate to lose that knowledge.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>

[REWRITTEN COMMIT: incomplete subtree]