Merge support::ErrorCode and galois::ErrorCode
It is confusing to have include files that only differ the base lib
because, in the final installed library, all the include directories
will be merged into a common include directory.

In general, try to reduce the number of namespaces in the code and
prefer unique names as much as possible due to how C++ resolves nested
namespaces.

[REWRITTEN COMMIT: incomplete subtree]