Apply consistent header guard
The particular pattern isn't important as long as it is consistent and
related to the filename. I picked the easiest one to automate, which is
based on the path to the file relative to the source root, followed up
by some cleanup to remove frequent components like "include" and "src".

There is a question about how to find header guards. I picked the simple
pattern of

  #ifndef ....
  #define ....

Parsing C, C++ comments to find the first preprocessor directive is an
alternative, but that seems more complicated than necessary.

There was only one case in the code base where this pattern was not a
header guard (libgpu/include/exclusive.h). In that case, I simply
modified the code to avoid my regex:

  #if !defined(....)
  #define ....

Another option would be to check if the ifdef macro matches the define
macro as a sign of guard. But a not uncommon error in header guards is
for those two macros to be different. Given the infrequency of valid
uses of this pattern, I opted to treat unequal macros as an error.

[REWRITTEN COMMIT: incomplete subtree]