CompilerQueryNode id into an int
For some reason, the compiler query node ID was a string that was pretty
much always casted to an int whenever it was used. This commit changes
the id type to unsigned int and removes the casts. It also adds an error
check to query graph construction to make sure that uninitialized id
(max value of uint32) doesn't pass.

[REWRITTEN COMMIT: incomplete subtree]