python: Support numpy dtype structs in Galois data and loops.
This adds struct support to InsertBag, LargeArray, and the
loops do_all and for_each, and OBIM metrics. This also adds
support for some additional primitive types (32-bit types).

To implement this, I refactored a lot of the wrapping
utilities and automation. I tried to make the use of these
more consistent as well so that the code is easier to
understand.

[REWRITTEN COMMIT: incomplete subtree]