Save property restrictions on nodes/edges
In Cypher, users can limit matches on matched nodes or edges by
specifying a property as well as what the property should be. For
example, the following node would only match if id was 123:

match (a {id:123})

This commit adds support for saving this information into a map for
later use in query graph IR construction. This will be done added in
a separate commit. Tests have been added to make sure parsed properties
are sane.

[REWRITTEN COMMIT: incomplete subtree]