Do not bind main thread by default
Binding the main thread is appropriate when running benchmark
applications when we can assume we have complete access to a machine.
But some current and future workloads may incorporate libgalois as a
library, and binding threads is an unexpected side-effect of using a
library.

Continue to bind worker threads by default. This is a little more
forgivable because this only affects parallel work spawned by the
libgalois itself.

There is a modest performance impact from not binding the main thread
because the NUMA node for TID 0 no longer corresponds to NUMA node 0.
However, the most common allocation policy for graphs and other large
data structures is UseNumaAlloc=false aka allocate interleaved, which
suggests our main performance concern is not overloading any particular
memory controller rather than aligning a thread's access to the closest
NUMA node.

[REWRITTEN COMMIT: incomplete subtree]