Save variable length edges into QueryGraph
CompilerQueryEdge gets two new fields to save variable length path
information, and this information gets propagated to the QueryGraph
object's edges as well. This will allow the execution engine to detect
variable length paths and react accordingly.

Disabled the old compiler code that used label strings to mark variable
length edges. It is much easier to just use a boolean on the Edge object
to do this. Added tests for the new variable length edges and left some
TODOs regarding various things, most notably a problem with the
finite repeating edge infrastructure not passing a test that it should
pass.

[REWRITTEN COMMIT: incomplete subtree]