Drop incoming edges (CSC) from LabeledGraph used in querying until we have support for distributed partitions with incoming edges (CSC). I expect to undo this commit once we have that support.
This change affects the queries that are supported. For example, personcityuniversity.q in sample_queries/two is not supported after this change. That query will be removed from ctest.

This change also affects the matching orders that can be supported. Each query vertex in the matching order is  expected to be connected to at least on vertex before it using only incoming edges.

This change is also likely to affect performance because searching/iterating the incoming edges is not supported.

[REWRITTEN COMMIT: incomplete subtree]