Fix compiler errors in serialize/deserialize
Compiler errors are fixed by using forward declaration
and std::enable_if.

Also fix is_memory_copyable: the definition used custom definition
of __is_trivially_copyable, which was compiler dependent.
However, this was incorrect as it returned true for
PODResizeableArray in gcc 9.3 but false for PODResizeableArray
in clang 10.0.1. The fix is to directly use std::is_trivially_copyable
(instead of custom definition) as we do not support non-compliant
compilers anymore.

Also simplify is_serialize (although it not used currently)

[REWRITTEN COMMIT: incomplete subtree]