Refactor to use op graph + fixes; inputs/output
Renames querytest to graph-query and changes it to use the OpGraph. A
hack was added to deal with distributed count for the time being as the
tables are not synchronized. Some new functions were added to
QueryServer and Table to facilitate this change.

Updates the test inputs to v4: some queries had to be changed binding
the variable from a path to a vertex because the OpGraph (and
technically even the older execution engine) did not support return
variables bound to paths, and this was causing tests to fail. v4 has
removed the queries and verify file: instead, they have been moved to
the graph-query directory so that a small change does not require a
completely new inputs tar to be created.

[REWRITTEN COMMIT: incomplete subtree]