python: Make PropertyGraph.get_*_property return Array if possible
This provides nearly 10x performance improvement in some cases.

This slightly changes the type of get_{edge,node}_property to
return an Array if the chunked array of the property only has one
chunk (the chunked array is returned otherwise). This is the most
common path. The APIs of ChunkedArray and Array are similar enough
that they are interchangable for most uses in python and in numba
compiled code. This change adds get_{edge,node}_property_chunked
to get the original chunked array if needed.

The reason for this change is that accesses to chunked arrays are
much slower due to the use of reference counted shared_ptrs and
the need to iterate over the chunks in several cases.

[REWRITTEN COMMIT: incomplete subtree]