Update clang-format
We don't plan to actively sync with ISS/Galois, so we don't need to
follow their code formatting. The number of active developers is small
enough that we can afford a couple more tweaks the code formatting
before it gets prohibitively expensive.

Compared to the previous style, there is considerably less horizontal
whitespace alignment, which should be helpful for avoiding merge
conflicts.

The formatting style is based off the Google style, with the following
notable exceptions:
- in places where the Google style allows some freedom, prefer a
  canonical output (e.g., AllowShortLoopsOnASingleLine: false,
  AlwaysBreakAfterReturnType: TopLevelDefinitions, DeriveXYZ: false)
- include ordering was adjusted to allow third-party inputs to be
  included by `#include <foo/bar.h>` rather than #include "foo/bar.h"

[REWRITTEN COMMIT: incomplete subtree]