Distribute graph for querying
This commit enables distributing the input graph
for running cypher queries.

The attributed graph used for querying is expected to be
pre-loaded into the query server. The attributed graph
requries the DistGraph to be constructed. The DistGraph
itself is destroyed after attributed graph is constructed.

The attributed graph stores distribution information as it
provides a partitioned view of the graph. All node identifiers
in its interface (both input and output) are global node IDs.

The graph-query app has been moved to distributed
sub-folder because distributed execution support is on
by default.

[REWRITTEN COMMIT: incomplete subtree]