Check for ifstream errors more consistently
While looking at libtsuba, I noticed that we could be a little more
precise in how we check errors from ifstreams.

Due to the ifstream API, we generally have to check the stream status
after we perform an operation (e.g., in.fail(), in.good(), in.bad(),
in.eof()).

It is possible to defer checking the stream status because the stream
state will propagated through operations but we cannot assume that any
stream output is correct until we check the stream status.

The typical pattern is:

  ifstream in;
  .... some operation with in ....
  if (!in) {  // equivalent to if (in.fail()) { ... }
    // an error happened
  } else {
    // end of file or in.good(), i.e., the last operation worked
  }

This avoids the confusingly named functions in.fail(), in.bad(),
in.good(), in.is_open(), and avoids the knowing eof is sometimes treated
like an error and sometimes not. It keeps the simple mental model
of "did my last operation work?".

I corrected ifstream uses in the more actively developed and used parts
of the code base: libgalois, libtsuba, dist, querying and tools.

[REWRITTEN COMMIT: incomplete subtree]