Access properties by type instead of position
Accessing properties by position was a transitional state. This commit
completes the transition to access properties by "name" as represented
as a C++ type. This is a more stable way to access fields.

The betweenness centrality and kcore applications were previously
excluded from the build, so I did not update them to access properties
by name. We will need better support for non-trivially constructable
types as properties in order to support them again.

Previously, the type `void` would be used to indicate that no node or
edge properties are present; however, the same idea can be represented
as `std::tuple<>`. This commit updates applications to use the latter
always. This simplifies the number of template tricks necessary to
support properties.

[REWRITTEN COMMIT: incomplete subtree]