tsuba: use name/version service to identify RDGs
This is a precursor to another PR coming soon which will connect to an
external name/version service.

Visible changes from to the graph engine: users of tuba should no longer
pass the name of a "meta file"; rather they should consider the
directory that the graph is stored in as the name of the graph. Doing
so, graphs can be updated in place and tsuba will manage the meta files
under the hood to ensure that latest version is read.

Partially to support some existing tests and partially to lay the
groundwork for loading old versions of the graph, I added new interfaces
to `Load` take a uri instead of a handle, these functions accept the
either graph names or meta files. My intention is to move towards a
world where holding a `handle` corresponds to holding a lock on the RDG.
So these interfaces could be used to read a graph that you intend to
write in a new location or read an old version of a graph that you
intend to write back as the latest version, allowing the implementation
of time travel.

There are a lot of churn-y clean up changes especially in RDG which I
thought were important, and definitely made reasoning about the
functionality changes easier.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>

[REWRITTEN COMMIT: incomplete subtree]