Reduce reliance on AtomicHelpers.h
AtomicHelpers provides some helpful functionality like atomic min and
atomic increment for floating point types (pre-C++20), but it also
contains reimplementations of STL functionality (std::fill,
std::transform) and basic C++ capabilities (variable assignment).

The main use of AtomicHelpers seems to be in supporting SyncStructures
which is itself mini-metaprogramming environment, and most of the
functionality in AtomicHelpers is not of general utility.

Remove various functions in AtomicHelpers in favor of their direct
implementation in SyncStructures.

Along the way, use fetch_add instead of compare_and_swap where we can.

[REWRITTEN COMMIT: incomplete subtree]