Reenable matrix completion and betweenness centrality
The matrix completion and betweenness centrality applications were
removed from the build because they require non-trivial data types in
their node data. However, by not building them, they have started to
drift with respect to the rest of the codebase. Reenable their
compilation but do not test them so that at least the code and its
dependencies is properly reflected through the rest of the codebase.

[REWRITTEN COMMIT: incomplete subtree]