name-service: connect to etcd if configured
The name service now operates in two modes: the first is the old
in-memory mode which technically provides the transactional guarantee
that we claim (with a single point of failure and the whole
"not-durable" thing) for testing.

In the second mode, it uses etcd to store changes. Etcd can run on
multiple nodes giving us a more palatable story about failures and also
logs to the local fs so there is some durability guarantee even if all
hosts crash (in practice it would log to elastic block store, whose
durability guarantees we should look into).

In the second mode, the name server instances that the katana graph
engine talks to have no application state so we could deploy an instance
of the name server on every host running graph computation (if the extra
network hop turns out to be a problem).

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>

[REWRITTEN COMMIT: incomplete subtree]