Create Distribution from PropertyFileGraph
As of now, there is no direct method of creating a Distribution without
invoking the Partitioner. This makes it difficult to create isolated
tests. Add a path for direct initialization of a Distribution using a
PropertyFileGraph.

While test-only interfaces create more maintenance burden, it seems
reasonable that a Distribution knows how to initialize itself
independent of the Partitioner.

Along the way, fix up the rvalue interface of Distribution and keep
partial initialization a Distribution between it and its friend, the
Partitioner.

[REWRITTEN COMMIT: incomplete subtree]