tsuba: don't check for error when trying to register a graph
I check to see if I can open a graph, then try to register it if it's
not found, but if concurrent tests are running it may have been
registered in the meantime and this register op will fail.

In this case we actually don't care if this test is the one to register
it or not so print the error if we get one but try to get it again from
the name-service anyway.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>

[REWRITTEN COMMIT: incomplete subtree]