Move page pool to galois::substrate
The line between the galois, galois::runtime and galois::substrate
namespaces is not very defined. Many data structures used by the
parallel runtime are also useful for library users as well (e.g.,
statistics, thread-local data structures). Conversely, the
galois::substrate namespace should be a firmer line because there is a
strict initialization order between basic memory and thread pools and
the data structures and executors built on top of them.

This commit and the following one move classes from galois::runtime to
either the galois or galois::substrate namespaces.

The page pool naturally belongs in galois::substrate. Along the way,
improve the interface to preallocation so that users do not need to
calculate page sizes themselves.

[REWRITTEN COMMIT: incomplete subtree]