Bug fix in cc and avoid dest dereferencing in GetData in apps
Shared-memory connected-components verification was failing (and occassionaly crashing)
due to the presence of 'const GNode& dest = *graph->GetEdgeDest(..)' as
dest was a garbage value most likely due to the reference to the temporary object
(although const is supposed to extend the lifetime of the temporary
object so not 100% sure).
This commit also makes call to 'GetEdgeDest' uniform across all the
lonestar apps and avoid unnecessary dereferencing of 'dest' when calling
'GetData(dest)'.