tsuba: use async write operations
For concurrency control some things are easier if writes are guaranteed
not to conflict. For instance, if I write meta_7 then die before I tell
the name service about it I wind up in a station where others can read
the stale `meta_7` when a real version 7 is committed.

To solve that problem I'm adding a random string to the end of meta
files. The version number will still serve to allow a new katana
instance to find the latest version of the graph if the directory is
moved, but the random tag will allow the name service to disambiguate
failed writes from successful ones.

In service of that I need to choose a tag, communicate it between hosts
then append it to meta and partition file names. In doing so I needed
a descriptor object that tracks the "tag" of the current write, and I
couldn't resist using this descriptor to collect futures and allow async
writes.

Tested by partitioning twitter40, blocked-cvc for 8 hosts. It used to
take ~30 seconds to write everthing out, now it takes ~15seconds. And
this is for a graph with very few properties, I expect the benefits in
more complex property graphs to be larger.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>

[REWRITTEN COMMIT: self-contained subtree]