tsuba: split the responsibilities of RDG
The new organization looks like this:

`RDGMeta` version information for RDGs, read/written from disk, points
to a set of RDGPartHeaders (implicitly through the file's name)

`RDGPartHeader` corresponds to the on-disk "partition file" that points
to all of the files needed to construct an RDG. It also contains
metadata about how the graph was partitioned.

`RDGCore` The property tables and topology for a partition. A bare bones
struct only intended for internal use.

`RDG` Wrapper around an `RDGCore` for use in applications.

`RDGSlice` Wrapper around an `RDGCore` for use in partitioning (doesn't
have attached partition information, does not support being written
out).

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>