Add network asserts
One possible but unconfirmed source of MPI timeouts is trying to call
MPI_Finalize while there are still unacknowledged MPI sends and recvs.
This commit adds a sanity check for one sufficient condition for this to
happen.

Along the way, clean up the network layer to remove unused support for
urgent messages, simplify the statistics reporting interface and push
down fields defined in a base class but only used in a subclass.

[REWRITTEN COMMIT: self-contained subtree]