Fix read-only view of string properties
The string value for each node was corrupted earlier
but the value was not complete garbage. The value
was a combination of substring values for other nodes.
This seems to be indicate that the bug was in our
interpretation of Arrow offsets. Instead of trying to
interpret internal Arrow data structures like offsets,
directly call appropriate Arrow functions.

Note that this bug existed only on distributed hosts
that did not read the properties direclty from disk.
This seems to indicate that the way we build the
string Arrow arrays after communication might be
messing up our interpretation of the Arrow offsets.